server/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
supabase
PyJWT
cryptography
requests
